feat: add new components including AppBar, InfoBar, and SharrowStatsBar; refactor imports for better organization
