+++ apps/web/src/app/layout.tsx
@@ -2,9 +2,9 @@ import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "@/components/ui/globals.css";
 import "@mysten/dapp-kit/dist/index.css";
-import AppBar from "@/components/AppBar";
 // import { ChallengeOverlay } from "@/components/ChallengeOverlay";
 import { ResultView } from "@/components/ResultView";
+import AppBar from "@/components/layout/AppBar";
 import { ConfettiEffect } from "@/components/ui/ConfettiEffect";
 import { Providers } from "@/providers/Providers";
 import { getMemeMetadataById, getRoundByNumber } from "@workspace/mockdata";
