+++ apps/web/src/components/layout/InfoBar.tsx
