+++ apps/web/src/components/stats/SharrowStatsBar.tsx
