+++ apps/web/src/hooks/index.ts
@@ -1,0 +1,6 @@
+// Transaction hooks
+export * from "./transactions";
+
+// Other hooks
+export { useChampionSwap } from "./useChampionSwap";
+export { useLenis } from "./useLenis";
